requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
pydantic==2.10.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
secure-smtplib==0.1.1
email-validator==2.2.0
python-dotenv==1.0.1
dnspython==2.7.0
openai==0.28.0
cryptography==44.0.0
pycryptodome==3.21.0
PyPDF2==3.0.1
PyJWT==2.10.1
bcrypt==3.2.0
passlib
sentence-transformers
